pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fueltracker"
version = "0.1.0"
description = "A Python package for tracking fuel prices and consumption"
authors = [
    {name = "Fuel Tracker Team", email = "team@fueltracker.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "requests>=2.32",
    "tenacity>=8.3",
    "pydantic>=2.7",
    "pyarrow>=16.0",
    "statsmodels>=0.14",
    "holidays>=0.56",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.9",
    "pytest>=7.0.0",
    "pre-commit>=3.0.0",
    "detect-secrets>=1.5.0",
]

[project.scripts]
fueltracker = "fueltracker.__main__:main"

[project.urls]
Homepage = "https://github.com/yourusername/fuel-tracker"
Repository = "https://github.com/yourusername/fuel-tracker"
Issues = "https://github.com/yourusername/fuel-tracker/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["fueltracker*"]

[tool.setuptools.package-data]
fueltracker = ["*.txt", "*.md", "*.yml", "*.yaml"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-q"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
# Enable: pycodestyle (E), pyflakes (F), bugbear (B), isort (I)
select = ["E", "F", "B", "I"]
# Put any ignores here (migrated from deprecated top-level):
# ignore = ["W293"]

# Optional: make fixes allowed when you pass --fix
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
# Tell isort what's "yours" so imports group correctly
known-first-party = ["fueltracker"]
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.format]
# Keep defaults similar to Black, LF endings
quote-style = "preserve"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"
